Batch remove action snapshots
This commit changes the way action snapshots are removed such that
they are deleted in bulk. Additionally it moves the archive process to
after the release process ends, preventing archiving from running
multiple times between a release.

It also removes the BIRTH action, which was included to ensure that an
action always had a parent. We don't need this. The only time an
action will not have a parent is during START.